Moving a couple things around.
