+++ packages/terra-data-grid/src/proptypes/sectionDataShape.js
