+++ packages/terra-data-grid/src/subcomponents/HeaderCell.module.scss
