+++ packages/terra-data-grid/src/subcomponents/ResizeHandle.module.scss
