+++ packages/terra-data-grid/src/subcomponents/Scrollbar.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import { DraggableCore } from 'react-draggable';
 
-import styles from './Scrollbar.scss';
+import styles from './Scrollbar.module.scss';
 
 const cx = classNames.bind(styles);
