+++ packages/terra-data-grid/src/subcomponents/SectionHeader.module.scss
